Support multiple references in a single cite statement.
For the CSL parser this is a bit of a hack, as the parser included, doesn't seem to support multiple references as input data.
Instead, we feed the parser each individual citation and the glue them together our selves. However, that does therefore not
support sorting or collating of references.